+++ src/app.controller.ts
@@ -1,12 +1,7 @@
-import { AuthService } from './auth/auth.service';
-import { NaverAuthGuard } from './auth/guard/naver-auth.guard';
-import { Controller, Get, Request, UseGuards } from '@nestjs/common';
-// import { AppService } from './app.service';
+import { Controller, Get } from '@nestjs/common';
 
-@Controller('api/v1')
+@Controller()
 export class AppController {
-  constructor(private readonly authService: AuthService) {}
-
   @Get('main')
   mainPageData() {
     return 'Main Page Router';
@@ -15,19 +10,5 @@ export class AppController {
   @Get('search')
   productSearch() {
     return 'Search Product';
-  }
-
-  @UseGuards(NaverAuthGuard)
-  @Get('auth/naver')
-  async login() {
-    return;
-  }
-
-  @UseGuards(NaverAuthGuard)
-  @Get('auth/naver/callback')
-  async callback(@Request() req): Promise<any> {
-    // 최종적으로 프론트로 보내주는 부분
-    return req.user;
-    // return this.authService.login(req.user);
   }
 }
